Reduce time to create a large test string
For the blk-archive testing we use a 8MiB string to generate files
that have the same content, different lengths, to maximize the de-dupe.
Previous implementation took ~2.2 seconds on my hardware to generate.
Replace with function that completes in ~0.006 seconds so we don't
induce a noticeable delay when interacting with dmtest.