Add ability to list and compare multiple tests
When --nr-run is not used with the list subcommand and there are
multiple runs of the test in the result set, it will print the number of
run in the listed state out of the total number of runs. Also, if there
are passing runs, the command will print the average duration of only
the passing runs. If there are no passing runs, the average duration of
all runs is printed.

When comparing result sets with multiple runs, the percentage of passed
runs is compared. The durations compared are the same as those shown
with the list command. The --nr-run option does not work with the
compare subcommand. There is no way to compare specific runs of two
result_sets to each other.

When filtering tests by state, the test matches if any of the runs
returned the desired state. To match tests where no run returned a
state, use "^STATE"

The log subcommand remains as before, where the first run is used if
--nr-run is not specified.