Add list-runs subcommand
Instead of having to manually run "list --run-nr" on each run to see
which ones passed and which ones failed, the list-runs subcommand shows
all the runs at once, along with their run number, so its easier to find
a specific run.  Using --run-state, the output can be filtered to only
show the runs in a certain state.