Sync metadata dev wipes
When dmest is run with the metadata device on a partition, another
partition may remain open after the dd command to wipe the device
completes, keeping the IO from getting automatically flushed back to the
device. This can lead to a situation where a thin-pool created
immediately afterwards will still see the old metadata. Call fsync after
wiping the metadata device to avoid this.